Separate standardization and convertion of an array, add unit tests

+++ tests/test_visualizer.py
@@ -1,5 +1,6 @@
 # pylint: disable = missing-class-docstring
 
+import numpy as np
 import pytest
 import torch
 from matplotlib.pyplot import Figure
@@ -11,8 +12,8 @@ def test_single_channel_visualization_selected_channel_should_raise_exception() 
     """Test if function raises ValueError when passing selected_channel
     not in range of color dimension of attributions matrix."""
     selected_channel: int = 4
-    attributions: torch.Tensor = torch.zeros((3, 2, 2))
-    transformed_img: torch.Tensor = torch.zeros((3, 2, 2))
+    attributions: torch.Tensor = torch.zeros((3, 2, 2), dtype=np.float)
+    transformed_img: torch.Tensor = torch.zeros((3, 2, 2), dtype=np.float)
     with pytest.raises(ValueError):
         _ = single_channel_visualization(
             attributions=attributions,
@@ -25,8 +26,8 @@ def test_single_channel_visualization_selected_channel_should_pass() -> None:
     """Test if function returns Figure object when passing selected_channel
     in range of color dimension of attributions matrix."""
     selected_channel: int = 0
-    attributions: torch.Tensor = torch.zeros((3, 2, 2))
-    transformed_img: torch.Tensor = torch.zeros((3, 2, 2))
+    attributions: torch.Tensor = torch.zeros((3, 2, 2), dtype=np.float)
+    transformed_img: torch.Tensor = torch.zeros((3, 2, 2), dtype=np.float)
     result = single_channel_visualization(
         attributions=attributions,
         transformed_img=transformed_img,
